type fixes and updates
