+++ packages/safe-apps-sdk/src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './sdk';
+export * from './rpc';
+export * from './gateway';
+export * from './messaging';
